moving InequalityTextField.kt, InequalityScreen.kt and InequalityVM.kt files from the `presentation` module to the `features:inequality:impl` module
